meetup-1/twitter-extractor/requirements.txt
elasticsearch==5.0.1
gender-detector==0.1.0
geolocation-python==0.2.2
nltk==3.2.1
oauthlib==2.0.1
python-dateutil==2.6.0
requests==2.12.1
requests-oauthlib==0.7.0
setuptools-scm==1.15.0
six==1.10.0
textblob==0.11.1
tweepy==3.5.0
urllib3==1.19.1
